Move support classes in test project
