+++ cs/TagsCloudVisualization_Tests/Picture.cs
