+++ cs/TagsCloudVisualization_Tests/SizesGenerator.cs
@@ -1,8 +1,7 @@
 ﻿using System;
-using System.Collections.Generic;
 using System.Drawing;
 
-namespace TagsCloudVisualization
+namespace TagsCloudVisualization_Tests
 {
     public class SizesGenerator
     {
